ANDROID: KVM: arm64: Handle guest FF-A unmap call
Unmap the guest buffers from the hypervisor when the guest invokes the
FFA_RXTX_UNMAP interface. Store the IPA of the buffers to be able to
unshare the memory from the hypervisor. Prevent the impersonation
if someone tries to unshare the buffers of another VM by checking the
handle from the vcpu.

Bug: 458241298
Signed-off-by: Sebastian Ene <sebastianene@google.com>
Signed-off-by: Fuad Tabba <tabba@google.com>
Change-Id: Id7241d8037215bbd246fb8282f3ddab20cec7dc7