ANDROID: KVM: arm64: iommu: Teardown guest domains
There is no guarantee that the guest will free it's domains before teardown,
one approach to solve this is to bruteforce through guest domain space,
and check the owner of each domain and free it.
This was simple to implement, but it won't scale properly and required some
hacks.

Another solution is to keep track of guest domains in a linked list and
keep it updated with alloc/free operations, and at teardown we free the
current domains in the list.

As the domains can be allocated without iommu context, we can't track
in an pviommu but it has to be done for a VM.

Also, this requires a new core IOMMU operation kvm_iommu_force_free_domain()
which force the domain refcount to zero before freeing it.

Bug: 357781595
Bug: 348382247
Bug: 236685427
Bug: 458241298
Change-Id: Iad8ce103fd6a5b6610f9773df1652e4bbc87d66f
Signed-off-by: Mostafa Saleh <smostafa@google.com>