ANDROID: KVM: arm64: THP support for pKVM guests
When a pKVM stage-2 guest fault happens, if it is within the remits of a
transparent huge page at stage-1, that entire page is pinned and we can
map the entire range of that page at stage-2. The hypervisor in return
installs a block mapping. This reduce TLB pressure and the number of
faults.

Bug: 458241298
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>
Signed-off-by: Fuad Tabba <tabba@google.com>
Change-Id: I8ba253134007ba31c39d243130384ac602c0f92e