ANDROID: KVM: arm64: Add __pkvm_host_map_guest_mmio()
When a guest touches a page, the host will map it, similarly to normal
memory. However, there is a separate hypercall for MMIO to avoid
increasing the cost of memory mapping which is the hot path.

Upon MMIO mapping, the hypervisor will atomically check that parts
of the group the device belonging to is already donated, and tag
those devices as owned by the VM.

Bug: 357781595
Bug: 348382247
Bug: 458241298
Change-Id: Ic4e3c048e2093da45a0717b239ab30cacc310e22
Signed-off-by: Mostafa Saleh <smostafa@google.com>