ANDROID: KVM: arm64: Mandate IO guard for guest physical MMIO
The host should only be allow to install MMIO for regions that the
guest have declared as MMIO with IO guard. That is critical to avoid
the host tricking the guest to access MMIO while it thinks it's normal
memory.
Replace the donation logic to check for IO guard first and remove it.

Bug: 357781595
Bug: 348382247
Bug: 458241298
Change-Id: I461a33cb9b8275a3dcfe88ce8501e11ae49dd954
Signed-off-by: Mostafa Saleh <smostafa@google.com>