ANDROID: KVM: arm64: Check supported VCPU features for protected VMs
Check at the host kernel which VCPU features are supported for pKVM. The
hypervisor performs a similar check, however, an unsupported feature
could be disabled silently. Ensure that userspace is aware if it tries
to set an unsupported feature for a protected VCPU.

Also, ensure that hyp and the host share the same code for checking
supported features for protected VMs.

This patch also fixes a bug where it was potentially setting
KVM_CAP_ARM_MTE, a VM capability, as a VCPU feature.

Bug: 458241298
Signed-off-by: Fuad Tabba <tabba@google.com>
Change-Id: Ice3e0364b13d4d1ad78926c62e38d35f744b7408